Minor updates to ReadMe and solver code
Added further details to sections within the ReadMe file. 
In CO2GraVISim_solver.f90 I've increased the length of the string for printing the plot times to stdout, and corrected a mistake in the coefficient of dh/dt.
In CO2GraVISim_model_description I've corrected a typo for a different part of that dh/dt coefficient.